requirements.txt
requests
pandas
pytest
azure-storage-blob
